django_PriMSA/requirements.txt
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
beautifulsoup4==4.4.1
biopython==1.66
bioservices==1.4.15
blinker==1.3
chardet==2.3.0
click==6.6
colorama==0.3.7
configobj==5.0.6
cryptography==1.2.3
cycler==0.9.0
Cython==0.25.2
Django==1.10.5
duplicity==0.7.6
easydev==0.9.31
enum34==1.1.2
Flask==0.10.1
gevent==1.2.1
greenlet==0.4.11
grequests==0.3.0
gyp==0.1
html5lib==0.999
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
liblarch==2.1.0
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==1.5.1
nose==1.3.7
numpy==1.11.1
numpydoc==0.6.0
panda==0.3.1
pexpect==4.2.1
Pillow==3.1.2
Pmw==1.3.2
primer3-py==0.5.1
ptyprocess==0.5.1
pyasn1==0.1.9
pychess==0.12.2
pygobject==3.20.0
pyinotify==0.9.6
pymol==1.7.2.1
pyOpenSSL==0.15.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2014.10
pyxdg==0.25
Reindent==0.1.1
reportlab==3.3.0
requests==2.13.0
requests-cache==0.4.13
scour==0.32
simplejson==3.8.1
six==1.10.0
sortedcontainers==1.4.4
sphinx-gallery==0.1.8
suds-jurko==0.6
unity-lens-photos==1.0
virtualenv==15.0.1
Werkzeug==0.10.4
wrapt==1.10.8
xmltodict==0.10.2
dj-database-url==0.3.0
django-postgrespool==0.3.0
gunicorn==19.3.0
psycopg2==2.6
django-toolbelt==0.0.1
static3==0.5.1
whitenoise==2.0.3
